moved isfullyqualified tests to net core app-specific file
